requirements.txt
arrow==0.12.1
base58==1.0.3
beautifulsoup4==4.7.0
bitcoin==1.1.42
certifi==2018.11.29
chardet==3.0.4
cycler==0.10.0
fpdf==1.7.2
freetype-py==2.0.0.post6
idna==2.8
moneywagon==1.21.10
numpy==1.15.4
Pillow==5.4.0
pycrypto==2.6.1
pyparsing==2.3.0
pypng==0.0.18
pysha3==1.0.2
python-dateutil==2.7.5
pytz==2018.7
qrcode==6.0
requests==2.21.0
scrypt==0.8.6
six==1.12.0
socketIO-client==0.7.2
soupsieve==1.6.1
tabulate==0.8.2
urllib3==1.24.1
websocket-client==0.54.0
